+++ components/lib/firebase-crud/read-data.ts
@@ -3,10 +3,10 @@ import { db } from "../firebase";
 
 export default async function getFirestoreData() {
   const querySnapshot = await getDocs(collection(db, "services"));
-  const servicedata = querySnapshot.docs.map((doc) => ({
+  const serviceData = querySnapshot.docs.map((doc) => ({
     ...doc.data(),
     id: doc.id,
   }));
 
-  return servicedata;
+  return serviceData;
 }
